pyproject.toml
[project]
name = "tg-sick-remote-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datetime>=5.5",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.8",
    "sqlalchemy>=2.0.36",
]
